pyproject.toml
[tool.poetry]
name = "vapor-steam"
version = "1.7.0"
description = "TUI program to check the ProtonDB compatibility of all the games of a Steam user."
authors = ["TabulateJarl8 <tabulatejarl8@gmail.com>"]
license = "GPLv3"
readme = "README.md"
packages = [{ include = "vapor" }]
homepage = "https://tabulate.tech/software/vapor"
repository = "https://github.com/TabulateJarl8/vapor"
keywords = ["steam", "protondb", "compatibility", "textual", "tui"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Typing :: Typed",
]
include = [{ path = "tests", format = "sdist" }]

[tool.poetry.dependencies]
python = "^3.10"
textual = "^7.2.0"
aiohttp = "^3.13.3"
platformdirs = [
  { version = ">=4.4.0,<4.5.0", python = "<3.10" },
  { version = "^4.5.0", python = ">=3.10" },
]
typing-extensions = "^4.15.0"
fastapi = "^0.135.1"
uvicorn = "^0.41.0"
python-dotenv = "^1.2.2"

[tool.poetry.scripts]
vapor = "vapor.__main__:main"

[tool.poetry.group.dev.dependencies]
textual-dev = "^1.8.0"
ruff = "^0.14.11"
pytest = [
  { version = "<9.0.0", python = "<3.10" },
  { version = "^9.0.0", python = ">=3.10" },
]
pytest-cov = "^7.0.0"
pytest-asyncio = [
  { version = "<1.3.0", python = "<3.10" },
  { version = "^1.3.0", python = ">=3.10" },
]
mypy = "^1.19.1"
pytest-xdist = "^3.8.0"
pytest-sugar = "^1.1.1"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.pyright]
reportUnusedCallResult = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
exclude_also = [
  # Don't complain about missing debug-only code:
  "def __repr__",
  "if self\\.debug",

  # Don't complain if tests don't hit defensive assertion code:
  "raise AssertionError",
  "raise NotImplementedError",

  # Don't complain if non-runnable code isn't run:
  "if 0:",
  "if __name__ == .__main__.:",

  # Don't complain about abstract methods, they aren't run:
  "@(abc\\.)?abstractmethod",

  # dont test functions that only make network calls
  "async def async_get",
]

omit = ["*/__main__.py"]

[tool.mypy]
strict = true
exclude = ["tests/"]

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
preview = true
extend-select = [
  # pycodestyle
  "E",
  # Warnings
  "W",
  # Pyflakes
  "F",
  # pydocstyle
  "D",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # pep8-naming
  "N",
  # flake8-bandit
  "S",
  # flake8-boolean-trap
  "FBT",
  # flake8-comprehensions
  "C4",
  # flake8-pie
  "PIE",
  # quote rules
  "Q001",
  "Q002",
  # flake8-raise
  "RSE",
  # flake8-return
  "RET",
  # flake8-unused-arguments
  "ARG",
  # flake8-use-pathlib
  "PTH",
  # eradicate
  "ERA",
  # perflint
  "PERF",
  # flake8-2020
  "YTT",
  # flake8-annotations
  "ANN",
  # flake8-async
  "ASYNC",
  # flake8-builtins
  "A",
  # flake8-commas
  "COM",
  # flake8-implicit-str-concat
  "ISC",
  # flake8-print
  "T20",
  # flake8-pytest-style
  "PT",
  # flake8-use-pathlib
  "PTH",
  # pylint
  "PL",
  # tryceratops
  "TRY",
  # refurb
  "FURB",
  # pydoclint
  "DOC",
  # ruff rules
  "RUF",
]

ignore = [
  # complains about tab indentation
  "W191",
  "D206",
  # adds a line break before a class docstring
  "D203",
  # puts the first line summary of a docstring on a different line than the """
  "D213",
  # tries to add a blank line after the last docstring section
  "D413",
  # yells at you if you use a bool typed function argument
  "FBT001",
  "FBT002",
  # yells at you for using try-except in a for loop
  "PERF203",
  # allow for the use of Any
  "ANN401",
  # false positives for overriding methods (i think)
  "PLR6301",
  # disable too many branches check
  "PLR0912",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN001", "ANN002", "PLC2701", "ARG002", "PLR2004", "DOC"]
"vapor/main.py" = ["DOC402"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
line-ending = "lf"

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true
order-by-type = true
